hostel_attendance/requirements.txt
deepface==0.0.93
opencv-python==4.9.0.80
numpy==1.26.4
scipy==1.13.0
python-dotenv==1.0.1
flask==3.0.3
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
sse-starlette==2.1.0
SQLAlchemy==2.0.30
